Move test to a different file
